Changed data context hierarchy to avoid reliance on hiding inherited member
We now have a BeforeSubmitChanges and CompleteSubmitChanges pair of
methods, so we no longer need to hide the SubmitChanges method.
Previously we needed to hide it so that when the base class (e.g.,
InMemoryContext) called SubmitChanges, the call would go to the base
class method instead of the derived class method (e.g., in
RaterooInMemoryDataContext), which had called it in the first place
through the BaseClassSubmitChanges method.